html/browsers/browsing-the-web/navigating-across-documents/javascript-url-query-fragment-components.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#javascript-protocol
  - id: webappapis.html#run-a-classic-script
    steps: [10]
  - id: https://url.spec.whatwg.org/#percent-decode
  - id: https://encoding.spec.whatwg.org/#utf-8-decode


description: |
  Creates iframes with src="javascript:..." and verifies
  the contents based on rules specified in steps 3-12

/-->
<title>javascript url with query and fragment components</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<!-- 9, ("nope" ? "yep" : "what";) is just a JavaScript expression to evaluate -->
<iframe id="a" src='javascript:"nope" ? "yep" : "what";'></iframe>
<!-- 4, %0a is treated as a byte and turned into "\n" (0x0A is \n), which means "ok" is the return value -->
<iframe id="b" src='javascript:"wrong"; // # %0a "ok";'></iframe>
<!-- 4, %25 is treated as a byte and turned into "%" (0x25 is %) -->
<iframe id="c" src='javascript:"%2525"'></iframe>

<script>
var test = async_test(undefined, {
  timeout: 1000
});
function check(id, expected) {
  assert_equals(
    document.getElementById(id).contentDocument.body.textContent,
    expected
  );
}
window.onload = test.step_func(function() {
  check("a", "yep");
  check("b", "ok");
  check("c", "%25");
  test.done();
});
</script>
